donations page, integrated nav
