+++ apps/frontend/src/app/page.tsx
@@ -1,5 +1,4 @@
 "use client";
-
 import NavBar from "./components/Navbar";
 
 export default function Home() {
@@ -7,7 +6,6 @@ export default function Home() {
     <div style={{ display: "flex", minHeight: "100vh" }}>
       <NavBar role="admin" />
       <main style={{ flex: 1, backgroundColor: "#f9fafb" }}>
-        {/* page content goes here */}
       </main>
     </div>
   );
